fix: display default text for current file
The image file doesn't have a 'name' property that the crispy-bootstrap5
template pack would normally use to add text to the anchor element. See
https://github.com/django-crispy-forms/crispy-bootstrap5/blob/0.6/crispy_bootstrap5/templates/bootstrap5/layout/field_file.html

The method I selected to resolve this was to copy the 'field_file'
template to my root templates directory so it is selected first, and
edit the template to use properties I specified.