refactor: rename role 'staff' to 'customer'
Rename the role to better represent the aims of the project.

During project planning, the scope was limited to an internal support
system where the role of a 'staff' member made sense for an
organization.

The project plan then developed to a more flexible software as a service
(SaaS) solution, but the role name 'staff' was overlooked and never
changed in the planning documentation. This change corrects that
oversight.