refactor: rename model 'Note' to 'Comment'
The model 'Note' to was renamed to 'Comment' after feedback that the
word 'Note' did not accurately represent the updates being added to a
ticket. A note would better represent an area to record private
observations whereas a comment better represents the public nature of
the messages passed back and forth between the customer and the
technician.

BREAKING CHANGE: The 'Note' model has been renamed, therefore references
to this model need to be refactored