fix: add common util function to ensure slug is an integer
The primary key (PK) is used as the slug throughout the project. If a
string was entered as the slug and not checked prior to the PK being
referenced in the view, an internal server error would occur. This
function coupled with its use in the 'dispatch' method throughout the
views, prevents any internal server errors being raised.