fix: validation errors when using summernote widget
The SummernoteWidget makes use of an iFrame which was being loaded with
attributes that caused HTML validation errors.

To correct this I made a copy of the 'widget_iframe' template to my root
templates directory so it is selected first. I then edited the template
to remove the attributes causing validation errors, replacing the width
with CSS.